// src/app/core/models/models.ts
export interface subMenus {
  url: string;
  separateRoute: boolean;
  menuValue: string;
  tittle: string;
  icon: string;
  showAsTab: boolean;
  showSubRoute: boolean;
  title: string;
  route?: string;
  base?: string;
  MenuItem: string;
}
// ...
export interface MenuItem {
  menuValue: string;
  showSubRoute: boolean;
  route: string;
  hasSubRoute: boolean;
  icon: string;
  base: string;
  url: string;
  hasSubRouteTwo2?:boolean;
}
// ...
export interface url {
  url: string;
}
// ...
export interface SideBar {
  showMyTab?: boolean;
  tittle: string;
  icon: string;
  showAsTab: boolean;
  separateRoute: boolean;
  materialicons?: string;
  menu: SideBarMenu[];
}
export interface SideBarMenu {
  showMyTab?: boolean;
  menuValue: string;
  menuValue1?: string;
  route?: string;
  hasSubRoute: boolean;
  showSubRoute: boolean;
  icon: string;
  base: string;
  materialicons: string;
  subMenus: SubMenu[];
}
export interface SubMenu {
  menuValue: string;
  menuValue1?: string;
  hasSubRouteTwo?:boolean;
  route?: string;
  base: string;
  base2?: string;
  base3?: string;
  base4?: string;
  base5?: string;
  base6?: string;
  base7?: string;
  base8?: string;
  subMenus?: SubMenu[]
}